moved init to cdp
