+++ Sources/Tracking/CustomerIO.swift
@@ -29,36 +29,6 @@ public extension CustomerIO {
      */
     @available(iOSApplicationExtension, unavailable)
     static func initialize(
-        writeKey: String,
-        configure configureHandler: ((inout SdkConfig, inout DataPipelineConfigOptions) -> Void)?
-    ) {
-        var sdkConfig = SdkConfig.Factory.create(siteId: "", apiKey: "", region: .US)
-        var cdpConfig = DataPipelineConfigOptions.Factory.create(writeKey: writeKey)
-
-        if let configureHandler = configureHandler {
-            configureHandler(&sdkConfig, &cdpConfig)
-        }
-
-        let newDiGraph = DIGraph(sdkConfig: sdkConfig)
-        let implementation = CustomerIOImplementation(diGraph: newDiGraph)
-
-        DataPipeline.initialize(moduleConfig: cdpConfig)
-
-        initializeSharedInstance(with: implementation, diGraph: newDiGraph, module: TrackingModuleHookProvider(), cleanupRepositoryImp: newDiGraph.cleanupRepository)
-
-        if sdkConfig.autoTrackScreenViews {
-            // Handle auto screen view tracking setup
-            // Note: setupAutoScreenviewTracking() needs to be implemented
-            shared.setupAutoScreenviewTracking()
-        }
-    }
-
-    /**
-     Initialize the shared `instance` of `CustomerIO`.
-     Call this function when your app launches, before using `CustomerIO.instance`.
-     */
-    @available(iOSApplicationExtension, unavailable)
-    static func initialize(
         siteId: String,
         apiKey: String,
         region: Region,
